requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
click==8.1.7
fastcore==1.7.5
fastlite==0.0.11
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.5.2
python-multipart==0.0.9
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlite-minutils==3.37.0.post3
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
